feat(seahaven-cli): restructure src directory
- Updated the `Cargo.toml` to change the path of the `truman` binary to `src/bin/truman/main.rs`.
- Introduced a new `lib.rs` file to define the Seahaven CLI library structure.
- Added multiple command modules under `src/bin/truman/cmd`, including `build`, `down`, `eject`, `init`, `pull`, `run`, `setup`, `system`, `up`, and `version`, enhancing the CLI's functionality.
- Implemented a common `--file` argument across commands for specifying the Seahaven setup file.
- Enhanced the `build` command to manage Docker image builds and added error handling for Docker dependencies.
- Improved the `run` command to execute tasks using the Just CLI, with support for dry runs and argument overrides.
