+++ seahaven-cli/src/bin/truman/cmd/build.rs
@@ -1,33 +1,34 @@
 use std::{fs::File, io::BufReader, path::PathBuf};
 
+use seahaven_cli::result::{Error, Result};
 use seahaven_docker::cmd::{DockerCmd, IntoCommand};
 
 use super::common::file_arg;
-use crate::result::{Error, Result};
-/// The `up` command name
-pub const CMD: &str = "up";
 
-/// Create the `up` command
+/// The `build` command name
+pub const CMD: &str = "build";
+
+/// Create the `build` command
 pub fn cmd() -> clap::Command {
     clap::command!(CMD)
-        .about("Start the development environment using docker compose")
-        .arg(file_arg().global(true))
+        .about("Build the development environment images using docker compose")
+        .arg(file_arg())
         .args([
-            clap::arg!(-d --detach "Detached mode: Run containers in the background")
-                .action(clap::ArgAction::SetTrue),
-            clap::arg!(--build "Build images before starting containers")
-                .action(clap::ArgAction::SetTrue),
             clap::arg!(--"dry-run" "Execute command in dry run mode")
                 .action(clap::ArgAction::SetTrue),
-            clap::arg!([SERVICE] ... "The services to start").action(clap::ArgAction::Append),
+            clap::arg!(--"build-arg" <KEY_VALUE> "Set build-time variables")
+                .action(clap::ArgAction::Append)
+                .value_parser(clap::builder::ValueParser::new(parse_build_arg)),
+            clap::arg!([SERVICE] ... "The services to build").action(clap::ArgAction::Append),
         ])
 }
 
-/// The `up` command implementation
+/// The `build` command implementation
 pub async fn run(matches: &clap::ArgMatches) -> Result<()> {
     // Check for requirements
     // - No min version for docker
     // - No min version for docker compose plugin (required)
+    // - No min version for buildx plugin (required)
     let docker_exe = seahaven_docker::exe::resolve_cli_executable()
         .map_err(|err| anyhow::anyhow!("Failed to resolve docker executable: {err}"))?;
 
@@ -42,6 +43,12 @@ pub async fn run(matches: &clap::ArgMatches) -> Result<()> {
     if docker_version.plugin_compose.is_none() {
         return Err(anyhow::anyhow!(
             "Failed to determine the docker compose plugin version. Is the docker compose plugin installed?"
+        )
+        .into());
+    }
+    if docker_version.plugin_buildx.is_none() {
+        return Err(anyhow::anyhow!(
+            "Failed to determine the buildx plugin version. Is the buildx plugin installed?"
         )
         .into());
     }
@@ -95,35 +102,127 @@ pub async fn run(matches: &clap::ArgMatches) -> Result<()> {
             .map_err(|err| anyhow::anyhow!("Failed to write docker-compose.yaml file: {err}"))?;
     }
 
+    // Process and sanitize the build args
+    let build_args = matches
+        .get_many::<(String, String)>("build-arg")
+        .unwrap_or_default()
+        .fold(
+            std::collections::BTreeMap::new(),
+            |mut acc, (key, value)| {
+                acc.insert(key, value);
+                acc
+            },
+        );
+
+    // Create and run the docker command
     let mut command = DockerCmd::with_executable(docker_exe)
         .compose()
         .with_file(content_file_path)
         .with_env_file(env_file_path)
         .with_plain_progress()
-        .up()
-        .with_build(matches.get_flag("build"))
-        .with_detach(matches.get_flag("detach"))
-        .with_services(matches.get_many::<String>("SERVICE").unwrap_or_default())
+        .build()
         .with_dry_run(matches.get_flag("dry-run"))
+        .with_build_args(build_args)
+        .with_services(matches.get_many::<String>("SERVICE").unwrap_or_default())
         .into_command();
 
-    tracing::debug!("Running command: {:?}", command.as_std());
+    tracing::debug!("command: {:?}", command.as_std());
 
     let mut child = command
         .kill_on_drop(true)
         .spawn()
-        .map_err(|err| anyhow::anyhow!("Failed to spawn docker compose up command: {err}"))?;
+        .map_err(|err| anyhow::anyhow!("Failed to spawn docker compose build command: {err}"))?;
 
     let status = child
         .wait()
         .await
-        .map_err(|err| anyhow::anyhow!("Failed to wait for docker compose up command: {err}"))?;
+        .map_err(|err| anyhow::anyhow!("Failed to wait for docker compose build command: {err}"))?;
     if !status.success() {
         return Err(
-            Error::new(anyhow::anyhow!("Docker compose up command failed"))
+            Error::new(anyhow::anyhow!("Docker compose build command failed"))
                 .with_code(status.code().unwrap_or(1)),
         );
     }
 
     Ok(())
+}
+
+/// Parse a build arg
+fn parse_build_arg(arg: &str) -> anyhow::Result<(String, String)> {
+    if let Some((key, value)) = arg.split_once('=') {
+        Ok((key.to_string(), value.to_string()))
+    } else {
+        Err(anyhow::anyhow!("Invalid build arg: {arg}"))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::parse_build_arg;
+
+    #[test]
+    fn parse_build_arg_with_valid_input() {
+        //* Given
+        let input = "FOO=bar";
+
+        //* When
+        let result = parse_build_arg(input);
+
+        //* Then
+        assert!(result.is_ok());
+        assert_eq!(result.unwrap(), ("FOO".to_string(), "bar".to_string()));
+    }
+
+    #[test]
+    fn parse_build_arg_with_empty_value() {
+        //* Given
+        let input = "FOO=";
+
+        //* When
+        let result = parse_build_arg(input);
+
+        //* Then
+        assert!(result.is_ok());
+        assert_eq!(result.unwrap(), ("FOO".to_string(), "".to_string()));
+    }
+
+    #[test]
+    fn parse_build_arg_with_multiple_equals() {
+        //* Given
+        let input = "FOO=bar=baz";
+
+        //* When
+        let result = parse_build_arg(input);
+
+        //* Then
+        let (key, value) = result.expect("Failed to parse build arg");
+        assert_eq!(key, "FOO");
+        assert_eq!(value, "bar=baz");
+    }
+
+    #[test]
+    fn parse_build_arg_with_no_equals() {
+        //* Given
+        let input = "FOO";
+
+        //* When
+        let result = parse_build_arg(input);
+
+        //* Then
+        let err = result.expect_err("Expected error");
+        assert!(err.to_string().contains("Invalid build arg"));
+    }
+
+    #[test]
+    fn parse_build_arg_with_empty_string() {
+        //* Given
+        let input = "";
+
+        //* When
+        let result = parse_build_arg(input);
+
+        //* Then
+        let err = result.expect_err("Expected error");
+        assert!(err.to_string().contains("Invalid build arg"));
+    }
 }
