+++ seahaven-cli/src/bin/truman/cmd/eject.rs
@@ -1,7 +1,8 @@
 use std::{fs::File, io::BufReader, path::PathBuf};
 
+use seahaven_cli::result::Result;
+
 use super::common::file_arg;
-use crate::result::Result;
 
 /// The `eject` command name
 pub const CMD: &str = "eject";
