+++ seahaven-cli/src/bin/truman/cmd/setup.rs
@@ -1,7 +1,9 @@
 use std::{fs::File, io::BufReader, path::PathBuf};
 
+use seahaven_cli::result::Result;
+
 use super::common::file_arg;
-use crate::result::Result;
+
 /// The `setup` command name
 pub const CMD: &str = "setup";
