+++ seahaven-cli/src/bin/truman/main.rs
@@ -1,5 +1,4 @@
 mod cmd;
-mod result;
 
 #[tokio::main(flavor = "current_thread")]
 pub async fn main() {
