+++ seahaven-cli/src/lib.rs
@@ -1,0 +1,3 @@
+//! # Seahaven CLI
+
+pub mod result;
